requirements.txt
amqp==5.0.6
asgiref==3.4.1
autopep8==1.5.7
beautifulsoup4==4.9.3
billiard==3.6.4.0
boto3==1.18.6
botocore==1.21.6
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cssutils==2.3.0
Django==3.2
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-filebrowser==3.14.1
django-filter==2.4.0
django-grappelli==2.15.1
django-inlinecss==0.3.0
django-js-asset==1.2.2
django-phonenumber-field==5.2.0
django-redis==5.0.0
django-storages==1.11.1
django-timezone-field==4.2.1
django-tinymce==3.3.0
Django-tinymce-filebrowser==0.2.1
djangorestframework==3.12.4
filebrowser-safe==0.5.0
flower==1.0.0
future==0.18.2
humanize==3.10.0
idna==3.2
jmespath==0.10.0
kombu==5.1.0
Markdown==3.3.4
phonenumberslite==8.12.28
Pillow==8.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pynliner==0.8.0
python-crontab==2.5.1
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
redis==3.5.3
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
sorl-thumbnail==12.7.0
soupsieve==2.2.1
sqlparse==0.4.1
stripe==2.60.0
toml==0.10.2
tornado==6.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
